Created two separate app files
