+++ packages/react-composites/tests/browser/chat/app/FakeAdapterApp.tsx
@@ -12,8 +12,8 @@ import { ChatAdapter, ChatComposite, createAzureCommunicationChatAdapterFromClie
 // eslint-disable-next-line no-restricted-imports
 import { IDS } from '../../common/constants';
 import { verifyParamExists } from '../../common/testAppUtils';
-import { FakeChatService } from './fake-back-end/ChatService';
-import { ChatAdapterModel } from './fixture';
+import { FakeChatService } from '../fake-adapter/fake-back-end/ChatService';
+import { ChatAdapterModel } from '../fake-adapter/fixture';
 
 const urlSearchParams = new URLSearchParams(window.location.search);
 const params = Object.fromEntries(urlSearchParams.entries());
@@ -21,7 +21,7 @@ const params = Object.fromEntries(urlSearchParams.entries());
 /**
  * App with chat composite using a fake Chat adapter
  */
-export const App = (): JSX.Element => {
+export const FakeAdapterApp = (): JSX.Element => {
   // Required params
   const fakeChatAdapterModel = JSON.parse(
     verifyParamExists(params.fakeChatAdapterModel, 'fakeChatAdapterModel')
